+++ packages/rum/src/domain/record/observers/observers.spec.ts
@@ -14,15 +14,8 @@ import {
   initFrustrationObserver,
   initInputObserver,
   initMouseInteractionObserver,
-  initMoveObserver,
 } from './observers'
-import type {
-  FrustrationCallback,
-  InputCallback,
-  MouseInteractionCallBack,
-  StyleSheetCallback,
-  MousemoveCallBack,
-} from './observers'
+import type { FrustrationCallback, InputCallback, MouseInteractionCallBack, StyleSheetCallback } from './observers'
 
 const DEFAULT_SHADOW_ROOT_CONTROLLER: ShadowRootsController = {
   flush: noop,
@@ -465,70 +458,5 @@ describe('initMouseInteractionObserver', () => {
       a.blur()
       expect(coordinatesComputed).toBeFalse()
     })
-  })
-})
-
-describe('initMoveObserver', () => {
-  let mouseMoveCallbackSpy: jasmine.Spy<MousemoveCallBack>
-  let stopObserver: () => void
-
-  beforeEach(() => {
-    if (isIE()) {
-      pending('IE not supported')
-    }
-
-    serializeDocument(document, DEFAULT_CONFIGURATION, {
-      shadowRootsController: DEFAULT_SHADOW_ROOT_CONTROLLER,
-      status: SerializationContextStatus.INITIAL_FULL_SNAPSHOT,
-      elementsScrollPositions: createElementsScrollPositions(),
-    })
-
-    mouseMoveCallbackSpy = jasmine.createSpy()
-    stopObserver = initMoveObserver(mouseMoveCallbackSpy)
-  })
-
-  afterEach(() => {
-    stopObserver()
-  })
-
-  it('should generate mouse move record', () => {
-    const event = createNewEvent('mousemove', { clientX: 1, clientY: 2 })
-    document.body.dispatchEvent(event)
-
-    expect(mouseMoveCallbackSpy).toHaveBeenCalledWith(
-      [
-        {
-          x: 1,
-          y: 2,
-          id: jasmine.any(Number),
-          timeOffset: 0,
-        },
-      ],
-      IncrementalSource.MouseMove
-    )
-  })
-
-  it('should generate touch move record', () => {
-    const event = createNewEvent('touchmove', { changedTouches: [{ clientX: 1, clientY: 2 }] })
-    document.body.dispatchEvent(event)
-
-    expect(mouseMoveCallbackSpy).toHaveBeenCalledWith(
-      [
-        {
-          x: 1,
-          y: 2,
-          id: jasmine.any(Number),
-          timeOffset: 0,
-        },
-      ],
-      IncrementalSource.TouchMove
-    )
-  })
-
-  it('should not generate mouse move record if x/y are missing', () => {
-    const mouseMove = createNewEvent('mousemove')
-    document.body.dispatchEvent(mouseMove)
-
-    expect(mouseMoveCallbackSpy).not.toHaveBeenCalled()
   })
 })
